qa: restrict revarsling to not older than 2 weeks. Move date filter to SQL query. Parametrized sendBrev func
