+++ src/main/kotlin/no/nav/syfo/service/FysiskBrevUtsendingService.kt
@@ -13,26 +13,15 @@ class FysiskBrevUtsendingService(
         uuid: String,
         journalpostId: String,
         distribusjonsType: DistibusjonsType,
+        tvingSentralPrint: Boolean = false,
     ) {
         val bestillingsId = journalpostdistribusjonConsumer.distribuerJournalpost(
             journalpostId,
             uuid,
-            distribusjonsType
+            distribusjonsType,
+            tvingSentralPrint = tvingSentralPrint
         ).bestillingsId
         log.info("Sendte til print, bestillingsId er $bestillingsId, varsel med UUID: $uuid")
-    }
-
-    suspend fun sendForcedBrev(
-        uuid: String, // aktivitetskravUuid
-        journalpostId: String,
-        distribusjonsType: DistibusjonsType,
-    ) {
-        val bestillingsId = journalpostdistribusjonConsumer.distribuerTvungetJournalpost(
-            journalpostId,
-            uuid,
-            distribusjonsType
-        ).bestillingsId
-        log.info("[FORCED PHYSICAL PRINT]: Sent forced letter to physical print, bestillingsId is $bestillingsId, varsel with UUID: $uuid")
     }
 }
